CiscoClientMERAKI: Fix progress reporting
Progress was showing 100% all the time.
This is due to progress only counting top-level endpoints
and there being only one (/organizations).

Having a single progress bar for all the nested child endpoints
would require redesigning the logic to not do separate recursive calls
for each parent's child.

Instead, ddd additional tasks to show nested progress bars
for fetching child endpoints,
which is a direct representation the logic,
at the cost of it looking more complicated for the user.

Pass the main CLI's rich console
to the rich progress instance
to make logging not clobber the multi-line progress.

Import the console from the cli package directly
instead of changing the CiscoClientController API
(which would require touching the other clients)
for now.